# FSync translog outside of the writers global lock