@mikemccand 2016-05-16
I think this change is potentially a biggish performance gain on highly concurrent CPUs with slowish IO.

I tested on 72 core box, on a spinning disk, and without the change (2 runs):

```
  67.3 M docs in 948.4 seconds
  67.3 M docs in 946.3 seconds
```
and 2 runs with the change:

```
  67.3 M docs in 858.1 seconds
  67.3 M docs in 884.7 seconds
```

I also ran the "random power loss tester" and there were no corruptions after 18 power loss events (at which point I hit disk full!).