@bleskes 2016-05-17
@mikemccand out of curiosity - how many concurrent threads did you use to index and how many shards are in the index?