@mikemccand 2016-05-17
>  how many concurrent threads did you use to index and how many shards are in the index?

6 shards, 12 client side threads, and I told ES to use 36 cores (though I did have a local mod to relax its current 32 core max).  I disabled auto-IO merge throttling, and ran with translog durability `async`.